Backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
asyncpg==0.29.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
pydantic==2.5.0
pgvector==0.2.4
google-generativeai==0.8.5
google-api-python-client==2.110.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
gspread==5.12.4
numpy==1.24.4
requests==2.31.0
aiofiles==23.2.1
websockets==12.0
langchain
openai
Pillow
redis==5.0.1
bcrypt